Move Arbitrary FunctionK instance
